Initial movement and fixes
