+++ web/src/components/status/ApplicationStatus.vue
@@ -114,7 +114,7 @@
 <script lang="ts">
 import { Component, Vue } from 'vue-property-decorator';
 import * as SurveyVue from "survey-vue";  
-import * as surveyEnv from "@/components/survey/question-types/survey-init";
+import * as surveyEnv from "@/components/survey/survey-init";
 import moment from 'moment-timezone';
 import {applicationInfoType} from "@/types/Application"
