+++ web/src/sandbox/components/SurveySandBox.vue
@@ -14,7 +14,7 @@
 <script lang="ts">
 import { Vue, Component, Prop } from "vue-property-decorator";
 import * as SurveyVue from "survey-vue";
-import * as SurveyInit from "@/components/survey/question-types/survey-init";
+import * as SurveyInit from "@/components/survey/survey-init";
 import SandboxSidebar from "./SandboxSidebar.vue";
 import Axios from "axios";
 import { addCustomTemplating } from "@/components/survey/survey-templating";
